==> BRICS/03 Scripts/01 BRICS _cleaned.ipynb <==
# BRICS cleaning data

#### This script contains

    01 Data Wrangling
    02 Quality and Consistency Checks
    03 Export of world_hapiness_cleaned1.csv

# Import libraries
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib
import os
import sklearn
from sklearn.model_selection import train_test_split 
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\6. Advanced Analytics and Dashboard Design'

# Import Data
df = pd.read_csv(os.path.join(path, 'BRICS', '02 Data', 'Original Data', 'Life_Expectancy_00_20.csv'))


## 01 Data wrangling

    Renaming columns
    

###  Renaming columns


# checking shape
df.shape

# checking column names
df.columns

# renaming columns
df.rename(columns={'Country' : 'country', 'Year' : 'year', 'People using at least basic drinking water services' : 'drinking_water', 'People practicing open defecation' : 'open_defecation', 'Continent' : 'continent', 'Least Developed' : 'least_developed', 'Life Expectancy' : 'life_expectancy', 'Population' : 'population', 'CO2 emissions' : 'co2_emissions', 'Health expenditure' : 'health_expenditure', 'Electric power consumption' : 'electric_power_consumption', 'Forest area' : 'forest_area', 'GDP per capita' : 'gdp_per_capita', 'Individuals using the Internet' : 'internet_users', 'Military expenditure' : 'military_exp', 'Obesity among adults' : 'obesity_among_adults', 'Beer consumption per capita' : 'beer_cons_per_capita' }, inplace = True)


# checking data types
df.dtypes


#### No need for changing data types nor Transposing data

## 02 Quality & Consistency Checks

    quality checks
    consistency checks

### quality checks

#### mixed type variables

# Überprüfung auf gemischte Datentypen
for col in df.columns.tolist():
    weird = (df[[col]].apply(lambda x: type(x.iloc[0])) != df[[col]].applymap(type)).any(axis=1)
    if len(df[weird]) > 0:
        print(col)

#### missing values


# Check for columns with missing values
df.isnull().sum()


# checking missing values by year
df.groupby(['year']).country.nunique()

#dataframe from 2016 to 2020
df_missing=df[df['year'].isin([2016,2017,2018,2019,2020])]

# list of countries that have no data
y=set(df['country'].unique())-set(df_missing['country'].unique())#
y

# delete coutries with missing values
df = df[~df['country'].isin(y)]


for country in set(df['country'].unique()):
    mean1 = df.loc[df["country"] == country, 'beer_cons_per_capita'].mean()
    mean2 = df.loc[df["country"] == country, 'electric_power_consumption'].mean()
    mean3 = df.loc[df["country"] == country, 'obesity_among_adults'].mean()
    mean4 = df.loc[df["country"] == country, 'health_expenditure'].mean()

    df.loc[df["country"] == country, 'beer_cons_per_capita'].fillna(mean1, inplace=True)
    df.loc[df["country"] == country, 'electric_power_consumption'].fillna(mean2, inplace=True)
    df.loc[df["country"] == country, 'obesity_among_adults'].fillna(mean3, inplace=True)
    df.loc[df["country"] == country, 'health_expenditure'].fillna(mean4, inplace=True)

df['country'].value_counts()

# Check for columns with missing values
df.isnull().sum()

# No missing values

#### duplicate values

# creating a subset with duplicates
df_dups = df[df.duplicated()]

# checking for values in the subset
df_dups

# No duplicates found

### Consistency checks

#### value counts for categorical columns 

df['country'].value_counts(dropna=False)

df['continent'].nunique()

df['continent'].value_counts(dropna=False)

df['least_developed'].value_counts(dropna=False)

df.columns

## 03 EXPORT world_hapiness_cleaned1.csv

# save the  dataframe as a csv in prepeared data folder
df.to_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_cleaned1.csv'))
