==> BRICS/03 Scripts/02 BRICS_joined.ipynb <==
# Joining happy data with API data - world_hapiness_joined

    01 Import Data via API
    02 Data Wrangling
    03 Join Data
    04 Export world_hapiness_joined.csv

import quandl
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
import statsmodels.api as sm # Using .api imports the public access version of statsmodels, which is a library that handles 
# statistical models.
import os
import warnings # This is a library that handles warnings.

warnings.filterwarnings("ignore") # Disable deprecation warnings that could indicate, for instance, a suspended library or 
# feature. These are more relevant to developers and very seldom to analysts.

plt.style.use('fivethirtyeight') # This is a styling option for how your plots will appear. More examples here:
# https://matplotlib.org/3.2.1/tutorials/introductory/customizing.html
# https://matplotlib.org/3.1.0/gallery/style_sheets/fivethirtyeight.html

path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\6. Advanced Analytics and Dashboard Design'

# Import Data world_hapiness
df_happy = pd.read_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_cleaned1.csv'))


df_happy.head()

from datetime import timedelta

df_happy.dtypes

## 01 Import Data via API

### Import Data

import nasdaqdatalink

nasdaqdatalink


# Configure API key 

nasdaqdatalink.ApiConfig.api_key = 'bp2zW4s1iBsxAXLXoVxY'

# Liste der Ländercodes, die du abrufen möchtest
countries = ['USA', 'BRA', 'CHN', 'IND', 'RUS', 'ZAF', 'FRA', 'DEU', 'CAN', 'JAP', 'ITA', 'GBR']  # Füge hier alle Länder hinzu, die du benötigst

# Liste der Indikatoren, die du abrufen möchtest
indicators = ['LE', 'GGXWDG_NGDP', 'GGXWDN', 'PPPSH', 'NGDPD', 'LP', 'BCA_NGDPD', 'BCA']  # Füge hier alle gewünschten Indikatoren hinzu

# LE - Employment, Millions
# NGDPD - GDP
# LP - Population
# GGXWDG_NGDP  - General Government Net Debt, % of GDP	 Staatsverschuldung
# GGXWDN - Debt
# BCA_NGDPD - Current Account Balance, % of GDP	 Außenhandelsbilanz
#  BCA-  Current Account Balance, USD Billions	




# DataFrame erstellen, um die Daten zu speichern
df= pd.DataFrame()

# Daten für jedes Land und jeden Indikator abrufen
for country in countries:
    for indicator in indicators:
        indicator_code = f'{country}_{indicator}'
        data = nasdaqdatalink.get_table('QDL/ODA', indicator=indicator_code)
        
        # Wenn Daten vorhanden sind und die Spalte 'date' vorhanden ist
        if not data.empty and 'date' in data.columns:
            # Daten auf den Zeitraum zwischen 2000 und 2020 filtern
            data = data[data['date'].between('1999-01-01', '2019-12-31')]
            
            # Füge die gefilterten Daten zum Ergebnis-DataFrame hinzu
            df = pd.concat([df, data], ignore_index=True)

# Zeige den aktualisierten DataFrame an
print(df)

# Splitte die Spalte 'indicator' am Unterstrich
df[['country', 'indicator1']] = df['indicator'].str.split('_', n=1, expand=True)

# Zeige den aktualisierten DataFrame an
print(df)

df['country'].value_counts()

df.columns

# Die Spalte 'indicator' vor dem Mergen entfernen
df.drop('indicator', axis=1, inplace=True)


### Create a pivot



# Pivot-Tabelle erstellen
pivot_df = df.pivot_table(index=['date', 'country'], columns='indicator1', values='value', aggfunc='first').reset_index()



# Zeige den aktualisierten DataFrame an
print(pivot_df)


pivot_df.shape


# LE - Employment, Millions
# NGDPD - GDP
# LP - Population
# GGXWDG_NGDP  - General Government Net Debt, % of GDP	 Staatsverschuldung
# GGXWDN - Debt
# BCA_NGDPD - Current Account Balance, % of GDP	 Außenhandelsbilanz
#  BCA-  Current Account Balance, USD Billions	

# renaming columns
pivot_df.rename(columns={'GGXWDG_NGDP' : 'debt_GDP', 'GGXWDN' : 'debt', 'LE' : 'employment', 'PPPSH' : 'share_of_world_GDP_ppp', 'LP':'population1', 'NGDPD': 'gdp', 'BCA_NGDPD':'account_balance_GDP', 'BCA':'account_balance'}, inplace = True)

pivot_df['country'].value_counts()

pivot_df.tail()

#### +1 day für pivot_df

# Einen Tag zu den Datumswerten in der 'date'-Spalte hinzufügen
pivot_df['date'] = pd.to_datetime(pivot_df['date']) + timedelta(days=1)

# Zeige den aktualisierten DataFrame an
print(pivot_df.head())



## 02 Data Wrangling

pivot_df.shape

pivot_df.columns

pivot_df.describe()

#### drop columns

# Check for columns with missing values
pivot_df.isnull().sum()


# Dropping column
pivot_df = pivot_df.drop(['debt'], axis=1)
pivot_df = pivot_df.drop(['employment'], axis=1)


### missing values

pivot_df['country'].value_counts()

missing_data = pivot_df[pivot_df.isnull().any(axis=1)]
missing_data

#  just elder data missing

df_happy.columns


df_happy.shape

df_happy.head()

### country names pivot

print(df_happy['country'].unique())
print(df_happy['year'].unique())
print(pivot_df['country'].unique())
print(pivot_df['date'].dt.year.unique())

print(pivot_df['country'].unique())


#### mapping der pivot

# Liste der Ländercodes und vollständigen Namen
country_mapping = {
    'USA': 'United States',
    'BRA': 'Brazil',
    'CHN': 'China',
    'IND': 'India',
    'RUS': 'Russian Federation',
    'ZAF': 'South Africa',
    'FRA': 'France',
    'DEU': 'Germany',
    'CAN': 'Canada',
    'JAP': 'Japan',
    'ITA': 'Italy',
    'GBR': 'United Kingdom'
}

# Setze die 'country'-Spalte in pivot_df zurück und setze die Ländercodes direkt
pivot_df['country'] = pivot_df['country'].apply(lambda x: country_mapping.get(x, x))



pivot_df['country'].value_counts()

df_happy['country'].value_counts()

### Create filtered df_happy

selected_countries = ['United States', 'Brazil', 'China', 'India', 'Russian Federation', 'South Africa', 'France', 'Germany', 'Canada', 'Japan', 'Italy', 'United Kingdom']

# Filtere df_happy basierend auf den ausgewählten Ländern
df_happy_filtered = df_happy[df_happy['country'].isin(selected_countries)]



df_happy['country'].value_counts()

df_happy_filtered['country'].value_counts()


# Schritt 3: Überprüfen der ausgewählten Länder in beiden DataFrames
print("Ausgewählte Länder in df_happy:", set(df_happy['country']))
print("Ausgewählte Länder in pivot_df:", set(pivot_df['country']))
print("Ausgewählte Länder in df_happy_filtered:", set(df_happy_filtered['country']))


df_happy_filtered.columns


#### check der typen

df_happy_filtered.dtypes


pivot_df.dtypes

pivot_df['date'] = pd.to_datetime(pivot_df['date'])
df_happy_filtered.rename(columns={'year': 'date'}, inplace=True)
df_happy_filtered['date'] = pd.to_datetime(df_happy_filtered['date'], format='%Y')

## 03 Join Data



# Führe den Join durch
joined_df = df_happy_filtered.set_index(['country', 'date']).join(pivot_df.set_index(['country', 'date']), how='inner')

# Setze die Indexspalten zurück, falls gewünscht
joined_df.reset_index(inplace=True)

print("Joined DataFrame:")
print(joined_df)

# Führe den Join erneut durch
#merged_df = pd.merge(df_happy_filtered, pivot_df, how='inner', on=['country', 'date'])
#print("Merged DataFrame:")
#print(merged_df)

# Überprüfe Spaltennamen und -reihenfolge
print("Spalten in df_happy_filtered:", df_happy_filtered.columns)
print("Spalten in pivot_df:", pivot_df.columns)


joined_df.shape


joined_df.head()

joined_df.describe()

joined_df.columns

## 04 EXPORT world_hapiness_joined.csv

# save the dataframe as a csv in prepeared data folder
joined_df.to_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_joined.csv'))
