==> BRICS/03 Scripts/03 BRICS_final.ipynb <==
# BRICS_final

#### This script contains

    1.data wrangling
    2.quality and consistency checks
    4.exploring relationships
    04 Regression

!pip install --upgrade bottleneck


# Import libraries
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib
import os
import sklearn
from sklearn.model_selection import train_test_split 
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score
import math


path = r'C:\Users\Lenovo\OneDrive\Dokumente\CareerFoundry\Data Immersion\6. Advanced Analytics and Dashboard Design'

# Import Data
df = pd.read_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_joined.csv'))


## 1. Data wrangling

    Renaming columns
    

###  Renaming columns


# checking shape
df.shape

# checking columns
df.head()

# checking for mixed data types
for col in df.columns.tolist():
  weird = (df[[col]].apply(lambda x: type(x.iloc[0]) != type(x), axis=1))
  if len (df[weird]) > 0:
    print (col)
    


# checking data types
df.dtypes


# Einen Tag von der 'date'-Spalte abziehen
df['date'] = pd.to_datetime(df['date']) - pd.Timedelta(days=1)

# Konvertiere die Spalte 'date' in ein DateTime-Objekt
df['date'] = pd.to_datetime(df['date'])

# Setze die Spalte 'date' als Index
df.set_index('date', inplace=True)

# Extrahiere das Jahr aus dem DateTime-Index und füge es als neue Spalte 'year' hinzu
df['year'] = df.index.year



df.describe(include='all')


df.columns

#### Changing data types with panda function


# Hier sind die Spalten "column1" und "column2" diejenigen, die du umwandeln möchtest

df['gdp'] = df['gdp'].apply(pd.to_numeric, errors='coerce')
df['life_expectancy'] = df['life_expectancy'].apply(pd.to_numeric, errors='coerce')
df['population'] = df['population'].apply(pd.to_numeric, errors='coerce')
df['co2_emissions'] = df['co2_emissions'].apply(pd.to_numeric, errors='coerce')
df['health_expenditure'] = df['health_expenditure'].apply(pd.to_numeric, errors='coerce')
df['electric_power_consumption'] = df['electric_power_consumption'].apply(pd.to_numeric, errors='coerce')
df['forest_area'] = df['forest_area'].apply(pd.to_numeric, errors='coerce')
df['gdp_per_capita'] = df['gdp_per_capita'].apply(pd.to_numeric, errors='coerce')
df['internet_users'] = df['internet_users'].apply(pd.to_numeric, errors='coerce')
df['military_exp'] = df['military_exp'].apply(pd.to_numeric, errors='coerce')
df['open_defecation'] = df['open_defecation'].apply(pd.to_numeric, errors='coerce')
df['drinking_water'] = df['drinking_water'].apply(pd.to_numeric, errors='coerce')
df['obesity_among_adults'] = df['obesity_among_adults'].apply(pd.to_numeric, errors='coerce')
df['beer_cons_per_capita'] = df['beer_cons_per_capita'].apply(pd.to_numeric, errors='coerce')
df['account_balance'] = df['account_balance'].apply(pd.to_numeric, errors='coerce')
df['account_balance_GDP'] = df['account_balance_GDP'].apply(pd.to_numeric, errors='coerce')
df['debt_GDP'] = df['debt_GDP'].apply(pd.to_numeric, errors='coerce')
df['population1'] = df['population1'].apply(pd.to_numeric, errors='coerce')
df['share_of_world_GDP_ppp'] = df['share_of_world_GDP_ppp'].apply(pd.to_numeric, errors='coerce')

# 'errors'='coerce' setzt alle nicht umwandelbaren Werte auf NaN

df['gdp'].head()

nan_counts_per_year = df.groupby('year').agg(lambda x: x.isna().sum().sum())



nan_counts_per_year_df = nan_counts_per_year.reset_index()
nan_counts_per_year_df
# betroffene spalten debt_GDP (2000 und 2001), obesity_among_adults(2017-20), beer_cons_per_capita (2020), electric_power_cinsumption

df.isnull().sum()

# drop columns
columns_to_drop = ['Unnamed: 0.1', 'Unnamed: 0', 'least_developed', 'electric_power_consumption', 'forest_area', 'population1' ]

# Verwende die drop-Methode, um die angegebenen Spalten zu entfernen
df.drop(columns=columns_to_drop, inplace=True)

df.info()


df['country'].value_counts()

# Identify the float64 columns
#float_columns = df.select_dtypes(include='float64').columns

# Replace dots with commas in float columns for display
#df[float_columns] = df[float_columns].applymap(lambda x: str(x).replace('.', ','))


## Export world_hapiness_final.csv!!! for Tableau with, instead of . 

# save the dataframe as a csv in prepeared data folder
#df.to_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_final.csv'))

## 2. Quality & Consistency Checks

    quality checks
    consistency checks

### quality checks

#### mixed type variables

# checking for mixed data types
for col in df.columns.tolist():
  weird = (df[[col]].apply(lambda x: type(x.iloc[0]) != type(x), axis=1))
  if len (df[weird]) > 0:
    print (col)
    



#### missing values


# Check for columns with missing values
df.isnull().sum()


# checking missing values by year
df.groupby(['year']).country.nunique()

### Impute missing values


# Impute missed values by the average
for i in set(df['country'].unique()):
    mean1 = df.loc[df["country"] == i, 'beer_cons_per_capita'].mean()
    mean3 = df.loc[df["country"] == i, 'obesity_among_adults'].mean()
    mean4 = df.loc[df["country"] == i, 'health_expenditure'].mean()
    df['beer_cons_per_capita'].fillna(mean1, inplace=True)
    df['obesity_among_adults'].fillna(mean3, inplace=True)
    df['health_expenditure'].fillna(mean4, inplace=True)

df.isnull().sum()

df['country'].value_counts()

# Check for columns with missing values
df.isnull().sum()

# 4 missing values in debt_GDP                  

#### duplicate values

# creating a subset with duplicates
df_dups = df[df.duplicated()]

# checking for values in the subset
df_dups

# No duplicates found

### Consistency checks

#### value counts for categorical columns 

df['country'].value_counts(dropna=False)

df.columns

## plausible?

df.shape

df['life_expectancy'].describe()

## 3. Adding new columns

    Adding new column
    creating subsets g7 and brics

df.columns

### health costs

# Annahme: df ist dein DataFrame

# Berechne die neuen Spaltenwerte
df['health_total'] = df['health_expenditure'] / 100 * df['gdp']
df['military_total'] = df['military_exp'] / 100 * df['gdp']
df['debt_total'] = df['debt_GDP'] / 100 * df['gdp']

# Anzeige des aktualisierten DataFrame
print(df.head())

df.describe()

#### adding brics flag

# creating new column and putting in values for brics and g7 countries

df['brics_flag'] = np.where(df['country'] == 'Brazil', 
                             'brics',
                             np.nan)
df['brics_flag'] = np.where(df['country'] == 'Russian Federation', 
                             'brics',
                             df['brics_flag'])
df['brics_flag'] = np.where(df['country'] == 'India', 
                             'brics',
                             df['brics_flag'])
df['brics_flag'] = np.where(df['country'] == 'China', 
                             'brics',
                             df['brics_flag'])
df['brics_flag'] = np.where(df['country'] == 'South Africa', 
                             'brics',
                             df['brics_flag'])

df['brics_flag'].value_counts()

#### adding g6 flag

# creating new column and putting in values for brics and g7 countries

df['g6_flag'] = np.where(df['country'] == 'United States', 
                             'g6',
                             np.nan)
df['g6_flag'] = np.where(df['country'] == 'United Kingdom', 
                             'g6',
                             df['g6_flag'])
df['g6_flag'] = np.where(df['country'] == 'France', 
                             'g6',
                             df['g6_flag'])
df['g6_flag'] = np.where(df['country'] == 'Italy', 
                             'g6',
                             df['g6_flag'])
df['g6_flag'] = np.where(df['country'] == 'Canada', 
                             'g6',
                             df['g6_flag'])
df['g6_flag'] = np.where(df['country'] == 'Germany', 
                             'g6',
                             df['g6_flag'])

df['g6_flag'].value_counts()

#### adding us_brics flag

# creating new column and putting in values for brics and g7 countries

df['us_brics_flag'] = np.where(df['country'] == 'Brazil', 
                             'brics',
                             np.nan)
df['us_brics_flag'] = np.where(df['country'] == 'Russian Federation', 
                             'brics',
                             df['us_brics_flag'])
df['us_brics_flag'] = np.where(df['country'] == 'India', 
                             'brics',
                             df['us_brics_flag'])
df['us_brics_flag'] = np.where(df['country'] == 'China', 
                             'brics',
                             df['us_brics_flag'])
df['us_brics_flag'] = np.where(df['country'] == 'South Africa', 
                             'brics',
                             df['us_brics_flag'])

df['us_brics_flag'] = np.where(df['country'] == 'United States', 
                             'us',
                             df['us_brics_flag'])



df['us_brics_flag'].value_counts()

df.info()

### creating subsets

#### subset brics

df_brics = df.loc[ df['brics_flag'].isin(['brics'])]


#### subset g6

df_g6 = df.loc[ df['g6_flag'].isin(['g6'])]

#### United States

df_us_brics = df.loc[df['us_brics_flag'].isin(['us']) | df['us_brics_flag'].isin(['brics'])]


df_us_brics.info()

#### G7

# creating g7 subset
df_g6 = df.loc[df['g6_flag'].isin(['g6'])]

# count values for g7 countries
df_g6['country'].value_counts(dropna=False)

#df_g6.describe()

##### Japanese data is missing!

#### BRICS

# creating brics subset
df_brics= df.loc[df['brics_flag'].isin(['brics'])]

# count values for brics countries
df_brics['country'].value_counts(dropna=False)

# BRICS Create a scatterplot using matplotlib for another look at how the chosen variables plot against each other.

df_brics.plot(x = 'military_exp', y='health_expenditure',style='o') # The style option creates a scatterplot; without it, we only have lines.
plt.title('military_exp and health_expenditure (BRICS)')  
plt.xlabel('military_exp')  
plt.ylabel('health_expenditure')  
plt.show()

# Reshape the variables into NumPy arrays and put them into separate objects.

X_brics = df_brics['military_exp'].values.reshape(-1,1)
y_brics = df_brics['health_expenditure'].values.reshape(-1,1)

# Split data into a training set and a test set.

X_train_brics, X_test_brics, y_train_brics, y_test_brics = train_test_split(X_brics, y_brics, test_size=0.3, random_state=0)

# Create objects that contain the model summary statistics.

rmse_g7 = mean_squared_error(y_test_g7, y_predicted_g7) # This is the mean squared error
r2_g7 = r2_score(y_test_g7, y_predicted_g7) # This is the R2 score. 

# Print the model summary statistics. This is where you evaluate the performance of the model.

print('Slope:' ,regression_g7.coef_)
print('Mean squared error: ', rmse_g7)
print('R2 score: ', r2_g7)

# Create a dataframe comparing the actual and predicted values of y.

data_g7 = pd.DataFrame({'Actual': y_test_g7.flatten(), 'Predicted': y_predicted_g7.flatten()})
data_g7.head(10)

# Predict.

y_predicted_train_g7 = regression_g7.predict(X_train_g7) # This is predicting X_train!

rmse2_g7 = mean_squared_error(y_train_g7, y_predicted_train_g7)
r22_g7 = r2_score(y_train_g7, y_predicted_train_g7)

print('Slope:' ,regression_g7.coef_)
print('Mean squared error: ', rmse2_g7)
print('R2 score: ', r22_g7)

# Visualizing the training set results.

plot_test_g7 = plt
plot_test_g7.scatter(X_train_g7, y_train_g7, color='green', s = 15)
plot_test_g7.plot(X_train_g7, y_predicted_train_g7, color='red', linewidth =3)
plot_test_g7.title('G7 military_exp vs health_expenditure  (Train set)')
plot_test_g7.xlabel('military_exp')
plot_test_g7.ylabel('health_expenditure')
plot_test_g7.show()

## Export world_hapiness_final_with_dots.csv

# save the  dataframe as a csv in prepeared data folder
df.to_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'world_hapiness_final_with_dots.csv'))


# save the  dataframe as a csv in prepeared data folder
df_us_brics.to_csv(os.path.join(path, 'BRICS', '02 Data', 'Prepared Data', 'df_us_brics.csv'))
